Modify token_generator.asm to match pseudocode in README.
Initial if statement checks to see if a new token should be generated, a token
is transmitted once the button is released (GenT pin goes high). Control drops
through to code below which will propagate a token coming in on the TI input.

Care must be taken to not generate multiple tokens in the system, always hold
the remove token button for a short period before adding a token.