Flip polarity of TI/TO lines in token generator firmware.
Same kind of changes as with token passer firmware. Swapping sets for clears
etc.

The change in token polarity is needed as the LVDS driver ICs used in the
interconnect idle high when no input is connected, we do not want a board
thinking it has a token coming in when things aren't connected properly.