Extend the token to 7us
The token was 5us previous and the polling loop took about 4us. This meant
that there was a worst case of 1us between the loop checking for the token
and the token finishing. Due to the fact the token is going out onto some
cabling with losses and so on, this 5us pulse was sometimes slightly shorter
so the token would disappear. With 7us I've yet to see a token disappear :D